Update structure for jetto submodule
