+++ duqtools/jetto/__init__.py
@@ -1,0 +1,9 @@
+"""This module contains tools for interfacing with jetto runs."""
+
+from ._namelist import read_namelist, patch_namelist, write_namelist
+
+__all__ = [
+'read_namelist',
+'patch_namelist',
+'write_namelist',
+]
